multiple stream msgs, arbitrary preceding characters
The previous solution assumed incorrectly that there was only message to
decode. It also incorrectly assumed that the preceding characters were
in nice neat 8-bit "words".

This escalated quickly. Fixing the problem was straight-forward. But, I
decided to add some testing to this, per your e-mail. In doing so, I
decided to leverage ES6 and ended up changing a lot in the process.

I think it supports the guidelines more closely now. Hope it works!