Buffer Length
To prevent an overflow of the buffer, you need to delete from the buffer every so often. I figured a two preamble-length buffer would suffice, but it didn't solve the "buffer around the boundary" problem, that is, a preamble who *mostly* sits in the buffer and has just one or two+ bites outside. If you completely delete the buffer, you still lose out, no matter how big. Made it so that now it only deletes the bits off the front of the buf that we're *sure* are not the beginning of a preamble.